UnkFunc58

+++ src/world/common/UnkFunc57.inc.c
@@ -63,7 +63,7 @@ ApiStatus N(UnkFunc57)(Evt* script, s32 isInitialCall) {
                 if (duration <= 0) {
                     duration = 1;
                 }
-                enemy->varTable[3] = fx_shape_spell(0, npc1->pos.x, npc1->pos.y + 14.0f, npc1->pos.z,
+                enemy->varTable[3] = (s32)fx_shape_spell(0, npc1->pos.x, npc1->pos.y + 14.0f, npc1->pos.z,
                                                     gPlayerStatusPtr->position.x,
                                                     gPlayerStatusPtr->position.y + 10.0f + 14.0f,
                                                     gPlayerStatusPtr->position.z, duration);
